Proposed wording change
In our lab at the top was a disclaimer, above the rending of this markdown (i don't see it in the markdown) it said - ..'Please use this resource group when creating the virtual machine'.   This confused me when I got to Task 2, as I thought we had to create a VM, and when I went to do this (bear in mind this is the tail end of the course when we are tired!) I went to the marketplace to make one and policy stopped me.  I was confused until someone point out the lab environment was a VM!   Suggested rewording for clarity